+++ exercicios CWI/Aula01/Exercicio001/src/Calculadora.java
@@ -1,18 +1,18 @@
 public class Calculadora {
 
-    public float soma(float num1, float num2){
+    public double soma(double num1, double num2){
         return num1 + num2;
     }
 
-    public float subtracao(float num1, float num2){
+    public double subtracao(double num1, double num2){
         return num1 - num2;
     }
 
-    public float multiplicacao(float num1, float num2){
+    public double multiplicacao(double num1, double num2){
         return num1 * num2;
     }
 
-    public float divisao(float num1, float num2){
+    public double divisao(double num1, double num2){
         return num1 / num2;
     }
